requirements.txt
# Modern UI Framework
customtkinter>=5.2.0

# Async HTTP Client
aiohttp>=3.8.0

# Async File Operations
aiofiles>=22.1.0

# Image Processing
Pillow>=9.0.0

# Additional modern UI enhancements
tkinter-tooltip>=2.0.0

# Color utilities for modern gradients
colorama>=0.4.0
